requirements.txt
requests
python-dotenv
python-gyazo